moved the navigation class together with navigation design file
