+++ lib/pages/navigation.dart
@@ -1,8 +1,49 @@
+import 'package:darth_runner/pages/profile.dart';
+import 'package:darth_runner/pages/run.dart';
+import 'package:darth_runner/pages/home.dart';
+import 'package:darth_runner/pages/stats_display.dart';
 import 'package:flutter/material.dart';
 import 'package:google_nav_bar/google_nav_bar.dart';
 
 
-//convert to statefulwidget when implementing navbar colour
+class NotHomePage extends StatefulWidget {
+  const NotHomePage({super.key});
+
+  @override
+  State<NotHomePage> createState() => _NotHomePageState();
+}
+
+class _NotHomePageState extends State<NotHomePage> {
+  int _selectedIndex = 0;
+
+  void navigateBottomBar(int index) {
+    setState(() {
+      _selectedIndex = index; 
+    });
+  }  
+
+  final List<Widget> _pages = [
+    const HomePage(),
+    const Run(),
+    const StatsDisplay(),
+    const Profile()
+  ];
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(   
+      body: _pages[_selectedIndex],
+      bottomNavigationBar:  MyBottomNavBar(
+        onTabChange: (index) => navigateBottomBar(index) ,
+        onTap: (index) => navigateBottomBar(index) ,
+      ),
+      
+    )
+    ;
+  }
+}
+
+// ignore: must_be_immutable
 class MyBottomNavBar extends StatelessWidget {
 
   void Function(int)? onTabChange;
